introduce IEntityToType for mapping LUIS EntityRecommendation to CLR Type
