+++ CSharp/Library/Microsoft.Bot.Builder.csproj
@@ -114,6 +114,8 @@
     <Compile Include="FormFlow\FieldReflector.cs" />
     <Compile Include="Fibers\Traits.cs" />
     <Compile Include="FormFlow\Localizer.cs" />
+    <Compile Include="Luis\CalendarPlus.cs" />
+    <Compile Include="Luis\EntityToType.cs" />
     <Compile Include="Luis\Extensions.cs" />
     <Compile Include="Luis\LuisModel.cs" />
     <Compile Include="Luis\Resolution.cs" />
